+++ titiler/stacapi/backend.py
@@ -1,117 +1,34 @@
 """titiler-stacapi custom Mosaic Backend and Custom STACReader."""
 
 import json
-from typing import Any, Dict, List, Optional, Tuple, Type
+from threading import Lock
+from typing import Any
 
 import attr
-import rasterio
 from cachetools import TTLCache, cached
 from cachetools.keys import hashkey
-from cogeo_mosaic.backends import BaseBackend
-from cogeo_mosaic.errors import NoAssetFoundError
-from cogeo_mosaic.mosaic import MosaicJSON
 from geojson_pydantic import Point, Polygon
 from geojson_pydantic.geometries import Geometry
 from morecantile import Tile, TileMatrixSet
-from pystac_client import ItemSearch
+from pystac_client import Client, ItemSearch
 from pystac_client.stac_api_io import StacApiIO
 from rasterio.crs import CRS
 from rasterio.warp import transform, transform_bounds
 from rio_tiler.constants import WEB_MERCATOR_TMS, WGS84_CRS
-from rio_tiler.errors import InvalidAssetName
-from rio_tiler.io import Reader
-from rio_tiler.io.base import BaseReader, MultiBaseReader
-from rio_tiler.models import ImageData
-from rio_tiler.mosaic import mosaic_reader
-from rio_tiler.types import AssetInfo, BBox
+from rio_tiler.mosaic.backend import BaseBackend, MosaicInfo
+from rio_tiler.types import BBox
+from rio_tiler.utils import CRS_to_uri
 from urllib3 import Retry
 
+from titiler.stacapi.dependencies import APIParams, Search
+from titiler.stacapi.reader import SimpleSTACReader
 from titiler.stacapi.settings import CacheSettings, RetrySettings, STACSettings
-from titiler.stacapi.utils import Timer
 
 cache_config = CacheSettings()
 retry_config = RetrySettings()
 stac_config = STACSettings()
 
-
-@attr.s
-class CustomSTACReader(MultiBaseReader):
-    """Simplified STAC Reader.
-
-    Inputs should be in form of:
-    {
-        "id": "IAMASTACITEM",
-        "collection": "mycollection",
-        "bbox": (0, 0, 10, 10),
-        "assets": {
-            "COG": {
-                "href": "https://somewhereovertherainbow.io/cog.tif"
-            }
-        }
-    }
-
-    """
-
-    input: Dict[str, Any] = attr.ib()
-    tms: TileMatrixSet = attr.ib(default=WEB_MERCATOR_TMS)
-    minzoom: int = attr.ib()
-    maxzoom: int = attr.ib()
-
-    reader: Type[BaseReader] = attr.ib(default=Reader)
-    reader_options: Dict = attr.ib(factory=dict)
-
-    ctx: Any = attr.ib(default=rasterio.Env)
-
-    def __attrs_post_init__(self) -> None:
-        """Set reader spatial infos and list of valid assets."""
-        self.bounds = self.input["bbox"]
-        self.crs = WGS84_CRS  # Per specification STAC items are in WGS84
-        self.assets = list(self.input["assets"])
-
-    @minzoom.default
-    def _minzoom(self):
-        return self.tms.minzoom
-
-    @maxzoom.default
-    def _maxzoom(self):
-        return self.tms.maxzoom
-
-    def _get_asset_info(self, asset: str) -> AssetInfo:
-        """Validate asset names and return asset's url.
-
-        Args:
-            asset (str): STAC asset name.
-
-        Returns:
-            str: STAC asset href.
-
-        """
-        if asset not in self.assets:
-            raise InvalidAssetName(
-                f"{asset} is not valid. Should be one of {self.assets}"
-            )
-
-        asset_info = self.input["assets"][asset]
-
-        url = asset_info["href"]
-        if alternate := stac_config.alternate_url:
-            url = asset_info["alternate"][alternate]["href"]
-
-        info = AssetInfo(url=url, env={})
-
-        if header_size := asset_info.get("file:header_size"):
-            info["env"]["GDAL_INGESTED_BYTES_AT_OPEN"] = header_size
-
-        if bands := asset_info.get("raster:bands"):
-            stats = [
-                (b["statistics"]["minimum"], b["statistics"]["maximum"])
-                for b in bands
-                if {"minimum", "maximum"}.issubset(b.get("statistics", {}))
-            ]
-            if len(stats) == len(bands):
-                info["dataset_statistics"] = stats
-
-        return info
+ttl_cache = TTLCache(maxsize=cache_config.maxsize, ttl=cache_config.ttl)  # type: ignore
 
 
 @attr.s
@@ -119,66 +36,43 @@ class STACAPIBackend(BaseBackend):
     """STACAPI Mosaic Backend."""
 
     # STAC API URL
-    url: str = attr.ib()
-    headers: Dict = attr.ib(factory=dict)
+    input: Search = attr.ib()
+    api_params: APIParams = attr.ib()
 
     # Because we are not using mosaicjson we are not limited to the WebMercator TMS
     tms: TileMatrixSet = attr.ib(default=WEB_MERCATOR_TMS)
-    minzoom: int = attr.ib()
-    maxzoom: int = attr.ib()
 
     # Use Custom STAC reader (outside init)
-    reader: Type[CustomSTACReader] = attr.ib(init=False, default=CustomSTACReader)
-    reader_options: Dict = attr.ib(factory=dict)
+    reader: type[SimpleSTACReader] = attr.ib(default=SimpleSTACReader)
+    reader_options: dict = attr.ib(factory=dict)
 
     # default values for bounds
     bounds: BBox = attr.ib(default=(-180, -90, 180, 90))
-
     crs: CRS = attr.ib(default=WGS84_CRS)
-    geographic_crs: CRS = attr.ib(default=WGS84_CRS)
-
-    input: str = attr.ib(init=False)
-    mosaic_def: MosaicJSON = attr.ib(init=False)
 
     _backend_name = "STACAPI"
 
-    def __attrs_post_init__(self) -> None:
+    def __attrs_post_init__(self):
         """Post Init."""
-        self.input = self.url
+        if bbox := self.input.get("bbox"):
+            self.bounds = tuple(bbox)
 
-        # Construct a FAKE mosaicJSON
-        # mosaic_def has to be defined.
-        # we set `tiles` to an empty list.
-        self.mosaic_def = MosaicJSON(
-            mosaicjson="0.0.3",
-            name=self.input,
-            bounds=self.bounds,
-            minzoom=self.minzoom,
-            maxzoom=self.maxzoom,
-            tiles={},
-        )
-
-    @minzoom.default
-    def _minzoom(self):
+    @property
+    def minzoom(self) -> int:
+        """Return minzoom."""
         return self.tms.minzoom
 
-    @maxzoom.default
-    def _maxzoom(self):
+    @property
+    def maxzoom(self) -> int:
+        """Return maxzoom."""
         return self.tms.maxzoom
 
-    def write(self, overwrite: bool = True) -> None:
-        """This method is not used but is required by the abstract class."""
-        pass
+    # in STACAPI backend assets are STAC Items as dict
+    def asset_name(self, asset: dict) -> str:
+        """Get asset name."""
+        return f"{asset['collection']}/{asset['id']}"
 
-    def update(self) -> None:
-        """We overwrite the default method."""
-        pass
-
-    def _read(self) -> MosaicJSON:
-        """This method is not used but is required by the abstract class."""
-        pass
-
-    def assets_for_tile(self, x: int, y: int, z: int, **kwargs: Any) -> List[Dict]:
+    def assets_for_tile(self, x: int, y: int, z: int, **kwargs: Any) -> list[dict]:
         """Retrieve assets for tile."""
         bbox = self.tms.bounds(Tile(x, y, z))
         return self.get_assets(Polygon.from_bounds(*bbox), **kwargs)
@@ -189,7 +83,7 @@ class STACAPIBackend(BaseBackend):
         lat: float,
         coord_crs: CRS = WGS84_CRS,
         **kwargs: Any,
-    ) -> List[Dict]:
+    ) -> list[dict]:
         """Retrieve assets for point."""
         if coord_crs != WGS84_CRS:
             xs, ys = transform(coord_crs, WGS84_CRS, [lng], [lat])
@@ -205,7 +99,7 @@ class STACAPIBackend(BaseBackend):
         ymax: float,
         coord_crs: CRS = WGS84_CRS,
         **kwargs: Any,
-    ) -> List[Dict]:
+    ) -> list[dict]:
         """Retrieve assets for bbox."""
         if coord_crs != WGS84_CRS:
             xmin, ymin, xmax, ymax = transform_bounds(
@@ -220,23 +114,33 @@ class STACAPIBackend(BaseBackend):
         return self.get_assets(Polygon.from_bounds(xmin, ymin, xmax, ymax), **kwargs)
 
     @cached(  # type: ignore
-        TTLCache(maxsize=cache_config.maxsize, ttl=cache_config.ttl),
-        key=lambda self, geom, search_query, **kwargs: hashkey(
-            self.url,
+        ttl_cache,
+        key=lambda self, geom, **kwargs: hashkey(
+            self.api_params["url"],
             str(geom),
-            json.dumps(search_query),
-            json.dumps(self.headers),
+            json.dumps(self.input),
+            json.dumps(self.api_params.get("headers", {})),
             **kwargs,
         ),
+        lock=Lock(),
     )
     def get_assets(
         self,
         geom: Geometry,
-        search_query: Optional[Dict] = None,
-        fields: Optional[List[str]] = None,
-    ) -> List[Dict]:
+        sortby: list[dict] | None = None,
+        limit: int | None = None,
+        max_items: int | None = None,
+        fields: list[str] | None = None,
+    ) -> list[dict]:
         """Find assets."""
-        search_query = search_query or {}
+
+        search_query = {
+            **self.input,
+            "method": "GET" if self.input.get("filter_expr") else "POST",
+            "sortby": sortby,
+            "limit": limit or 10,
+            "max_items": max_items or 100,
+        }
         fields = fields or ["assets", "id", "bbox", "collection"]
 
         stac_api_io = StacApiIO(
@@ -244,7 +148,7 @@ class STACAPIBackend(BaseBackend):
                 total=retry_config.retry,
                 backoff_factor=retry_config.retry_factor,
             ),
-            headers=self.headers,
+            headers=self.api_params.get("headers", {}),
         )
 
         params = {
@@ -255,86 +159,41 @@ class STACAPIBackend(BaseBackend):
         params.pop("bbox", None)
 
         results = ItemSearch(
-            f"{self.url}/search",
-            stac_io=stac_api_io,
-            **params,
+            f"{self.api_params['url']}/search", stac_io=stac_api_io, **params
         )
         return list(results.items_as_dicts())
 
-    @property
-    def _quadkeys(self) -> List[str]:
-        return []
+    @cached(  # type: ignore
+        ttl_cache,
+        key=lambda self: hashkey(
+            self.api_params["url"],
+            json.dumps(self.input),
+            json.dumps(self.api_params.get("headers", {})),
+        ),
+        lock=Lock(),
+    )
+    def info(self) -> MosaicInfo:  # type: ignore
+        """Mosaic info."""
+        renders = {}
+        bounds = self.bounds
+        crs = self.crs
 
-    def tile(
-        self,
-        tile_x: int,
-        tile_y: int,
-        tile_z: int,
-        search_query: Optional[Dict] = None,
-        **kwargs: Any,
-    ) -> Tuple[ImageData, List[str]]:
-        """Get Tile from multiple observation."""
-        timings = []
+        if collections := self.input.get("collections", []):
+            if len(collections) == 1:
+                stac_api_io = StacApiIO(
+                    max_retries=Retry(
+                        total=retry_config.retry,
+                        backoff_factor=retry_config.retry_factor,
+                    ),
+                    headers=self.api_params.get("headers", {}),
+                )
+                client = Client.open(f"{self.api_params['url']}", stac_io=stac_api_io)
+                collection = client.get_collection(collections[0])
+                if collection.extent.spatial:
+                    bounds = tuple(collection.extent.spatial.bboxes[0])
+                    crs = WGS84_CRS
+                renders = collection.extra_fields.get("renders", {})
 
-        with Timer() as t:
-            mosaic_assets = self.assets_for_tile(
-                tile_x,
-                tile_y,
-                tile_z,
-                search_query=search_query,
-            )
-
-        timings.append(("search", round(t.elapsed * 1000, 2)))
-
-        if not mosaic_assets:
-            raise NoAssetFoundError(
-                f"No assets found for tile {tile_z}-{tile_x}-{tile_y}"
-            )
-
-        def _reader(
-            item: Dict[str, Any], x: int, y: int, z: int, **kwargs: Any
-        ) -> ImageData:
-            with self.reader(item, tms=self.tms, **self.reader_options) as src_dst:
-                return src_dst.tile(x, y, z, **kwargs)
-
-        with Timer() as t:
-            img, used_assets = mosaic_reader(
-                mosaic_assets, _reader, tile_x, tile_y, tile_z, **kwargs
-            )
-
-        timings.append(("mosaicking", round(t.elapsed * 1000, 2)))
-        img.metadata = {**img.metadata, "timings": timings}
-        return img, used_assets
-
-    def point(
-        self,
-        lon: float,
-        lat: float,
-        coord_crs: CRS = WGS84_CRS,
-        search_query: Optional[Dict] = None,
-        **kwargs: Any,
-    ) -> List:
-        """Get Point value from multiple observation."""
-        raise NotImplementedError
-
-    def part(
-        self,
-        bbox: BBox,
-        dst_crs: Optional[CRS] = None,
-        bounds_crs: CRS = WGS84_CRS,
-        search_query: Optional[Dict] = None,
-        **kwargs: Any,
-    ) -> Tuple[ImageData, List[str]]:
-        """Create an Image from multiple items for a bbox."""
-        raise NotImplementedError
-
-    def feature(
-        self,
-        shape: Dict,
-        shape_crs: CRS = WGS84_CRS,
-        max_size: int = 1024,
-        search_query: Optional[Dict] = None,
-        **kwargs: Any,
-    ) -> Tuple[ImageData, List[str]]:
-        """Create an Image from multiple items for a GeoJSON feature."""
-        raise NotImplementedError
+        return MosaicInfo(
+            bounds=bounds, crs=CRS_to_uri(crs) or crs.to_wkt(), renders=renders
+        )